feet(PlaylistEntity): changed the logic playlist<->media type control (see details)
I decided to merge specialized playlists back to one general playlist. If it was left as before, specialized bridge entities would be necessary, which would cause the application to be harder to maintain and potentially expand.
So, I decided to add a media type flag into general playlist entity. The type of media entity will be checked with corresponding playlist type flag in the BL layer.